+++ frontend/src/pages/settings/Settings.tsx
@@ -28,6 +28,7 @@ import { useOpenWizard } from "../../components/wizard/SetupWizard";
 import { TabIDs } from "../../store/locationState";
 import {
   type Settings as ApiSettings,
+  Smart_mode,
   useGetApiSettingsQuery,
   usePutApiSettingsMutation,
 } from "../../store/sratApi";
@@ -94,6 +95,7 @@ export function Settings() {
   const methods = useForm({
     mode: "onBlur",
     values: globalConfig as ApiSettings,
+    defaultValues: { smart_mode: Smart_mode.Legacy },
     disabled: readOnly,
   });
   const { handleSubmit, reset, formState } = methods;